+++ metagenomescope/tests/layout/test_layout.py
@@ -68,8 +68,8 @@ def test_layout_to_solid_dot():
         ui_config.LAYOUT_DOT,
         {ui_config.LAYOUT_DOT: {"ranksep": 3}},
     )
-    w = lay.width / layout_config.PIXELS_PER_INCH
-    h = lay.height / layout_config.PIXELS_PER_INCH
+    w = lay.width / layout_config.POINTS_PER_INCH
+    h = lay.height / layout_config.POINTS_PER_INCH
     assert lay.to_solid_dot() == (
         f"  {cc.unique_id} [width={w},height={h},shape=rectangle,"
         f'label="cc{cc.cc_num}_id{cc.unique_id}"];\n'